refactor rendering and main

+++ src/consts.rs
@@ -11,7 +11,7 @@ pub const PATH_TRACE_RR: f64 = 0.2;
 
 /// How many shadow rays per vertex in path tracer? Preferably square for
 /// jittered sampler.
-pub const SHADOW_SPLITS: usize = 1;
+pub const SHADOW_SPLITS: u32 = 1;
 
 /*
  * TEXTURES
